Moved config from data_mappers to misc
