+++ lib/g910_gkey_mapper.py
@@ -4,8 +4,7 @@ import fcntl
 import os
 import signal
 import time
-import uinput
-from lib.data_mappers.config import Config
+from lib.misc.config import Config
 from lib.usb_device import USBDevice
 from lib.keyboard import Keyboard
 from lib.misc import notify, memory_leds
